Split statement tests into multiple files
